move mergeYoRcContent to lib
