+++ jdl/index.ts
@@ -1,4 +1,4 @@
 export * from './jdl-importer.js';
 export * from './parsing/api.js';
 export * from './jhipster/index.js';
-export * from './exporters/config.js';
+export * from '../lib/utils/yo-rc.js';
